static_root/requirements (copy).txt
BeautifulSoup==3.2.1
Brlapi==0.5.6
Django==1.4.1
Fabric==1.3.2
GnuPGInterface==0.3.2
Mako==0.5.0
Markdown==2.3
MarkupSafe==0.15
Mezzanine==1.4.10
MySQL-python==1.2.3
PAM==0.4.2
Pillow==2.4.0
PyBluez==0.18
SQLAlchemy==0.7.4
Scrapy==0.18.4
South==0.7.6
Twisted==13.1.0
Twisted-Core==11.1.0
Twisted-Names==11.1.0
Twisted-Web==11.1.0
adium-theme-ubuntu==0.3.2
apt-xapian-index==0.44
apturl==0.5.1ubuntu3
argparse==1.2.1
bleach==1.2.2
chardet==2.0.1
command-not-found==0.2.44
configglue==1.0
configobj==4.7.2
cssselect==0.9.1
debtagshw==0.1
decorator==3.3.2
defer==1.0.2
dirspec==3.0.0
distribute==0.7.3
dj-database-url==0.3.0
dj-static==0.0.5
django-admin-bootstrapped==0.4.3
django-email-as-username==1.6.5
django-filter==0.7
django-haystack==2.1.0
django-registration==1.0
django-tagging==0.3.1
django-taggit==0.12
django-tastypie==0.9.16
django-toolbelt==0.0.1
django-voting==0.2
djangorestframework==2.2.5
duplicity==0.6.18
filebrowser-safe==0.2.28
grappelli-safe==0.2.22
gunicorn==18.0
html5lib==0.95
httplib2==0.7.2
ipython==0.12.1
jockey==0.9.7
keyring==0.9.2
language-selector==0.1
launchpadlib==1.9.12
lazr.restfulclient==0.12.0
lazr.uri==1.0.3
louis==2.3.0
lxml==3.2.3
matplotlib==1.1.1rc
mimeparse==0.1.3
nose==1.1.2
numpy==1.6.1
nvidia-common==0.0.0
oauth==1.0.1
oauthlib==0.5.1
onboard==0.97.1
oneconf==0.2.8.1
paramiko==1.7.7.1
pexpect==2.3
piston-mini-client==0.7.2
protobuf==2.4.1
psycopg2==2.4.6
pyOpenSSL==0.13.1
pycrypto==2.4.1
pycups==1.9.61
pycurl==7.19.0
pygeoip==0.2.6
pyinotify==0.9.2
pylibmc==1.2.3
pyparsing==1.5.2
pyquery==1.2.8
pyserial==2.5
pysmbc==1.0.13
pysolr==3.0.5
python-apt==0.8.3ubuntu7
python-dateutil==1.5
python-debian==0.1.21ubuntu1
python-virtkey==0.60.0
pytz==2011k
pyxdg==0.19
queuelib==1.0
reportlab==2.5
requests==1.2.3
requests-oauthlib==0.3.2
rhythmbox-ubuntuone==3.0.0
selenium==2.33.0
sessioninstaller==0.0.0
signals==0.0.2
simplegeneric==0.7
simplejson==2.3.2
six==1.6.1
software-center-aptd-plugins==0.0.0
static==1.0.2
system-service==0.1.6
threadpool==1.2.7
ufw==0.31.1-1
unattended-upgrades==0.1
unity-lens-video==0.3.5
unity-scope-video-remote==0.3.5
usb-creator==0.2.23
vboxapi==1.0
w3lib==1.3
wadllib==1.3.0
wsgiref==0.1.2
wxPython-common==2.8.12.1
xdiagnose==2.5.2ubuntu0.1
xkit==0.0.0
xlrd==0.9.0
zope.interface==4.0.5
